Add PM table support for Rembrandt

+++ lib/nb_smu_ops.h
@@ -111,7 +111,7 @@ smu_t get_smu(nb_t nb, int smu_type);
 void free_smu(smu_t smu);
 u32 smu_service_req(smu_t smu ,u32 id ,smu_service_args_t *args);
 
-mem_obj_t init_mem_obj(u32 physAddr);
+mem_obj_t init_mem_obj(uintptr_t physAddr);
 
 void free_mem_obj(mem_obj_t obj);
